Deduplicate code and apply spotless style
